rename db.py -> db_fixtures.py
