+++ .secrets.baseline
@@ -163,10 +163,10 @@
         "is_secret": false
       }
     ],
-    "pytest_indexd/db.py": [
+    "pytest_indexd/fixtures/db_fixtures.py": [
       {
         "type": "Secret Keyword",
-        "filename": "pytest_indexd/db.py",
+        "filename": "pytest_indexd/fixtures/db_fixtures.py",
         "hashed_secret": "d033e22ae348aeb5660fc2140aec35850c4da997",
         "is_verified": false,
         "line_number": 78,
@@ -208,5 +208,5 @@
       }
     ]
   },
-  "generated_at": "2023-01-10T18:48:07Z"
+  "generated_at": "2023-01-10T19:45:44Z"
 }
