+++ pytest_indexd/plugin.py
@@ -1,4 +1,5 @@
-from pytest_indexd.db import (
+from pytest_indexd.fixtures.client_fixture import indexd_client
+from pytest_indexd.fixtures.db_fixtures import (
     alias_driver,
     auth_driver,
     create_indexd_tables,
@@ -7,6 +8,5 @@ from pytest_indexd.db import (
     pg_url,
     postgresql_server_indexd,
 )
-from pytest_indexd.fixtures.client_fixture import indexd_client
 from pytest_indexd.fixtures.loader_fixture import indexd_loader
 from pytest_indexd.fixtures.server_fixture import indexd_server
